wakeup-server/requirements.txt
fhir.resources==7.1.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JSON==0.4.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.6
